Making video clip by input seeking
